tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.put.null.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.put.null</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.put.null</h1>
<p class="desc">putImageData() with null imagedata throws TypeError</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("putImageData() with null imagedata throws TypeError");
_addTest(function(canvas, ctx) {

assert_throws(new TypeError(), function() { ctx.putImageData(null, 0, 0); });


});
</script>
